Updating fork to latest state

+++ xyz-hub-service/src/main/java/com/here/xyz/hub/auth/FeatureAuthorization.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017-2019 HERE Europe B.V.
+ * Copyright (C) 2017-2025 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,20 +19,23 @@
 
 package com.here.xyz.hub.auth;
 
+import static com.here.xyz.hub.auth.XyzHubActionMatrix.DELETE_FEATURES;
+import static com.here.xyz.hub.auth.XyzHubActionMatrix.UPDATE_FEATURES;
+import static com.here.xyz.util.service.BaseHttpServerVerticle.getJWT;
 import static io.netty.handler.codec.http.HttpResponseStatus.FORBIDDEN;
 
 import com.here.xyz.events.GetFeaturesByBBoxEvent;
 import com.here.xyz.events.GetFeaturesByGeometryEvent;
-import com.here.xyz.hub.rest.Api;
-import com.here.xyz.hub.rest.HttpException;
+import com.here.xyz.hub.connectors.models.Space;
 import com.here.xyz.hub.task.FeatureTask;
 import com.here.xyz.hub.task.FeatureTask.ConditionalOperation;
-import com.here.xyz.hub.task.FeatureTask.DeleteOperation;
 import com.here.xyz.hub.task.FeatureTask.GeometryQuery;
-import com.here.xyz.hub.task.ModifyOp.Entry;
-import com.here.xyz.hub.task.ModifyOp.IfExists;
 import com.here.xyz.hub.task.TaskPipeline.Callback;
-import com.here.xyz.models.geojson.implementation.Feature;
+import com.here.xyz.models.hub.jwt.ActionMatrix;
+import com.here.xyz.models.hub.jwt.AttributeMap;
+import com.here.xyz.util.service.HttpException;
+import com.here.xyz.util.service.logging.LogUtil;
+import io.vertx.ext.web.RoutingContext;
 
 public class FeatureAuthorization extends Authorization {
 
@@ -40,25 +43,41 @@ public class FeatureAuthorization extends Authorization {
   public static <X extends FeatureTask> void authorize(X task, Callback<X> callback) {
     if (task instanceof ConditionalOperation) {
       authorizeConditionalOp((ConditionalOperation) task, (Callback<ConditionalOperation>) callback);
-    } else if (task instanceof DeleteOperation) {
-      authorizeDeleteOperation((DeleteOperation) task, (Callback<DeleteOperation>) callback);
     } else {
       authorizeReadQuery(task, callback);
     }
+  }
+
+  public static void authorizeWrite(RoutingContext context, Space space, boolean isDelete) throws HttpException {
+    if(getJWT(context).skipAuth)
+      return;
+
+    final ActionMatrix tokenRights = getXyzHubMatrix(getJWT(context));
+    final XyzHubActionMatrix requestRights = new XyzHubActionMatrix();
+
+    //CREATE & UPDATE == WRITE
+    if (!isDelete && !requestRights.containsKey(UPDATE_FEATURES))
+      requestRights.updateFeatures(XyzHubAttributeMap.forValues(space.getOwner(), space.getId(), space.getPackages()));
+    //DELETE
+    else if (isDelete && !requestRights.containsKey(DELETE_FEATURES))
+      requestRights.deleteFeatures(XyzHubAttributeMap.forValues(space.getOwner(), space.getId(), space.getPackages()));
+
+    evaluateRights(LogUtil.getMarker(context), requestRights, tokenRights);
   }
 
   /**
    * Authorizes a query operation.
    */
   private static <X extends FeatureTask> void authorizeReadQuery(X task, Callback<X> callback) {
+
     //Check if anonymous token is being used
     if (task.getJwt().anonymous) {
       callback.exception(new HttpException(FORBIDDEN, "Accessing features isn't possible with an anonymous token."));
       return;
     }
 
-    if (task.space.isShared()) {
-      //User is trying to read a shared space this is allowed for any authenticated user
+    // Skip authorization if the space is shared or the service is running without authorization.
+    if (task.space.isShared() || task.getJwt().skipAuth) {
       callback.call(task);
       return;
     }
@@ -77,52 +96,35 @@ public class FeatureAuthorization extends Authorization {
         requestRights.readFeatures(XyzHubAttributeMap.forValues( ((GeometryQuery)task).refSpace.getOwner(), ((GeometryQuery)task).refSpaceId , ((GeometryQuery)task).refSpace.getPackages()));
     }
 
-    evaluateRights(requestRights, task.getJwt().getXyzHubMatrix(), task, callback);
+
+    evaluateRights(requestRights, getXyzHubMatrix(task.getJwt()), task, callback);
   }
 
   /**
    * Authorizes a conditional operation.
    */
   private static void authorizeConditionalOp(ConditionalOperation task, Callback<ConditionalOperation> callback) {
-    JWTPayload jwt = Api.Context.getJWT(task.context);
+    if (task.getJwt().skipAuth) {
+      callback.call(task);
+      return;
+    }
 
-    final ActionMatrix tokenRights = jwt.getXyzHubMatrix();
+    final ActionMatrix tokenRights = getXyzHubMatrix(getJWT(task.context));
     final XyzHubActionMatrix requestRights = new XyzHubActionMatrix();
 
-    task.modifyOp.entries.forEach(entry -> addAttributeMapForEntry(task, requestRights, entry));
+    //READ
+    if (!requestRights.containsKey(XyzHubActionMatrix.READ_FEATURES) && task.modifyOp.isRead())
+      requestRights.readFeatures(XyzHubAttributeMap.forValues(task.space.getOwner(), task.space.getId(), task.space.getPackages()));
+    //CREATE
+    else if (!requestRights.containsKey(XyzHubActionMatrix.CREATE_FEATURES) && task.modifyOp.isCreate())
+      requestRights.createFeatures(XyzHubAttributeMap.forValues(task.space.getOwner(), task.space.getId(), task.space.getPackages()));
+    //UPDATE
+    else if (!requestRights.containsKey(UPDATE_FEATURES) && task.modifyOp.isUpdate())
+      requestRights.updateFeatures(XyzHubAttributeMap.forValues(task.space.getOwner(), task.space.getId(), task.space.getPackages()));
+    //DELETE
+    else if (!requestRights.containsKey(DELETE_FEATURES) && task.modifyOp.isDelete())
+      requestRights.deleteFeatures(XyzHubAttributeMap.forValues(task.space.getOwner(), task.space.getId(), task.space.getPackages()));
 
     evaluateRights(requestRights, tokenRights, task, callback);
-  }
-
-  /**
-   * Authorizes a delete operation.
-   */
-  private static void authorizeDeleteOperation(DeleteOperation task, Callback<DeleteOperation> callback) {
-    final XyzHubActionMatrix requestRights = new XyzHubActionMatrix();
-    requestRights.deleteFeatures(XyzHubAttributeMap.forValues(task.space.getOwner(), task.space.getId(), task.space.getPackages()));
-
-    evaluateRights(requestRights, task.getJwt().getXyzHubMatrix(), task, callback);
-  }
-
-  private static void addAttributeMapForEntry(ConditionalOperation task, XyzHubActionMatrix requestRights,
-      Entry<Feature, Feature, Feature> entry) {
-
-    // READ
-    if (!requestRights.containsKey(XyzHubActionMatrix.READ_FEATURES) && (entry.head != null && IfExists.RETAIN
-        .equals(task.modifyOp.ifExists))) {
-      requestRights.readFeatures(XyzHubAttributeMap.forValues(task.space.getOwner(), task.space.getId(), task.space.getPackages()));
-    }
-    // CREATE
-    else if (!requestRights.containsKey(XyzHubActionMatrix.CREATE_FEATURES) && (entry.head == null && entry.result != null)) {
-      requestRights.createFeatures(XyzHubAttributeMap.forValues(task.space.getOwner(), task.space.getId(), task.space.getPackages()));
-    }
-    // UPDATE
-    else if (!requestRights.containsKey(XyzHubActionMatrix.UPDATE_FEATURES) && (entry.head != null && entry.result != null)) {
-      requestRights.updateFeatures(XyzHubAttributeMap.forValues(task.space.getOwner(), task.space.getId(), task.space.getPackages()));
-    }
-    // DELETE
-    else if (!requestRights.containsKey(XyzHubActionMatrix.DELETE_FEATURES) && entry.head != null && entry.result == null) {
-      requestRights.deleteFeatures(XyzHubAttributeMap.forValues(task.space.getOwner(), task.space.getId(), task.space.getPackages()));
-    }
   }
 }
